Make Money use a BigDecimal instead of double to improve accuracy.

+++ src/main/java/Money.java
@@ -1,12 +1,14 @@
+import java.math.BigDecimal;
+
 public class Money {
 
-    private double amount;
+    private BigDecimal amount;
 
-    public Money(double amount) {
+    public Money(BigDecimal amount) {
         this.amount = amount;
     }
 
     public double toDouble() {
-        return this.amount;
+        return this.amount.doubleValue();
     }
 }
